requirements.txt
Django==5.0.6
django-simple-captcha==0.5.14
django-ratelimit==3.1.0
Pillow
gunicorn==21.2.0